Plain vanilla description in the thrown error.
If there is a parsing error, the thrown Error object contains the
message which has the format:

  Error: Line 1: some error description

As suggested by Ricardo "mrdoob" Cabello, there should be a way to
grab the description only (without the prefix of "Error" and the
location) other than using regular expression to filter the message.

This is now possible using the 'description' property, e.g.

 try { esprima.parse('1 + 2 = 3') }
 catch (e) { console.log(e.description) }

http://code.google.com/p/esprima/issues/detail?id=389